proj as int now

+++ Core/DbConfig.cs
@@ -18,7 +18,7 @@
         // whether to create event-queue tables/indexes
         public bool queue;
 
-        public void ReadData(IDataInput i, ushort proj = 0)
+        public void ReadData(IDataInput i, int proj = 0)
         {
             i.Get(nameof(host), ref host);
             i.Get(nameof(port), ref port);
@@ -28,7 +28,7 @@
             i.Get(nameof(queue), ref queue);
         }
 
-        public void WriteData<R>(IDataOutput<R> o, ushort proj = 0) where R : IDataOutput<R>
+        public void WriteData<R>(IDataOutput<R> o, int proj = 0) where R : IDataOutput<R>
         {
             o.Put(nameof(host), host);
             o.Put(nameof(port), port);
